+++ src/main/java/ac/dnd/dodal/common/interceptor/UserIdArgumentResolver.java
@@ -2,7 +2,7 @@ package ac.dnd.dodal.common.interceptor;
 
 import ac.dnd.dodal.common.annotation.UserId;
 import ac.dnd.dodal.common.exception.UnauthorizedException;
-import ac.dnd.dodal.core.config.security.enums.SecurityExceptionCode;
+import ac.dnd.dodal.core.security.enums.SecurityExceptionCode;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
