+++ src/main/java/ac/dnd/dodal/common/util/JwtUtil.java
@@ -1,4 +1,4 @@
-package ac.dnd.dodal.core.config.security.util;
+package ac.dnd.dodal.common.util;
 
 import ac.dnd.dodal.common.constant.Constants;
 import ac.dnd.dodal.domain.user.enums.UserRole;
