+++ src/main/java/ac/dnd/dodal/core/security/info/JwtUserInfo.java
@@ -1,4 +1,4 @@
-package ac.dnd.dodal.core.config.security.info;
+package ac.dnd.dodal.core.security.info;
 
 import ac.dnd.dodal.domain.user.enums.UserRole;
 import lombok.Builder;
